Refactor to avoid app/instance-initializers/ for Ember 1.11 support.
Due to an issue with ember-load-initializers we cannot have a file in
`app/instance-initializers/` when running against Ember < 1.12.

That is fixed upstream in ember-load-initializers#0.1.7, but is not
commonly in use yet.

---

For now, move this to the `addon/` dir and register it manually during
the main initializer.
